==> 1-FeatSelec_1.1.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

#read file
ic50 = pd.read_csv("ic50_target.csv")
mordred = pd.read_csv("mordred_desc.csv")
padel = pd.read_csv("padel_desc.csv")

#drop column no in ic50_target
ic50.drop(columns=['no'], inplace=True)
ic50.head()

#drop column name in mordred_target
mordred.drop(columns=['name'], inplace=True)
mordred.head()

#drop column Name in mordred_target
padel.drop(columns=['Name'], inplace=True)
padel.head()

#membuat kolom label class dimana jika data di IC50 < 20 maka 1, selain itu 0 ( > 70)
ic50['class']=np.where(ic50['IC50 (uM)']< 20,1,0)
ic50.head()

#concat data mordred, padel & IC50 target
concat_df = pd.concat([padel,mordred,ic50], axis=1, join='outer')
concat_df.head()

#drop column Name in mordred_target
concat_df.drop(columns=['IC50 (uM)'], inplace=True)
concat_df.head()

#mengecek jumlah Nan pada tiap kolom
concat_df.isnull().sum()

#hapus kolom yang terdapat Nan
concat_df_clean = concat_df.dropna(axis=1)

#mengecek jumlah Nan pada tiap kolom yang sudah di clean
concat_df_clean.isnull().sum()

#memisahkan data train dan test dengan perbandingan 8:2
train, test = train_test_split(concat_df_clean, test_size=0.20, random_state=42, shuffle=True)

#mengetahui jumlah baris dan kolom test set
test.shape

#mengetahui jumlah baris dan kolom train set
train.shape

#ubah jadi file.csv
train.to_csv(r'train.csv', header=True, index=False)
test.to_csv(r'test.csv', header=True, index=False)
