==> 3-NonTuning_SVC_5_AdaBoost_PredictModel.ipynb <==
import pandas as pd
import pickle
import numpy as np
from sklearn.metrics import roc_curve, auc
from sklearn.svm import SVC
import matplotlib.pyplot as plt
from sklearn.ensemble import AdaBoostClassifier
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import classification_report
from sklearn.metrics import confusion_matrix, accuracy_score, precision_score, recall_score, f1_score, roc_auc_score, matthews_corrcoef

train = pd.read_csv('train.csv')
X_train = train.iloc[:,:-1]
y_train = train.iloc[:,-1]
train.shape

X_train = X_train[['SM1_Dzs','AATS1dv','C2SP2.1','AATSC0v','ATSC2m']]
X_train.head()

test = pd.read_csv('test.csv')
X_test = test.loc[:,['SM1_Dzs','AATS1dv','C2SP2.1','AATSC0v','ATSC2m']]
y_test = test.iloc[:,-1]
X_test.shape

#define model (parameter)
boost = AdaBoostClassifier(random_state=3)

#fit model
boost.fit(X_train, y_train)

# predicting the train set results
y_pred_train = boost.predict(X_train)

# predicting the test set results
y_pred_test = boost.predict(X_test)

y_proba_train = boost.predict_proba(X_train)

y_proba_test = boost.predict_proba(X_test)

y_proba_train = y_proba_train[:,1]

y_proba_test = y_proba_test[:,1]

#y_proba_train

#y_proba_test

AUC_train = roc_auc_score(y_train, y_proba_train)
AUC_train

AUC_test= roc_auc_score(y_test, y_proba_test)
AUC_test

fpr1, tpr1, _ = roc_curve(y_train, y_proba_train)
fpr2, tpr2, _ = roc_curve(y_test, y_proba_test)

#Evaluating the Algorithm train set

print('confusion matrix:')
print(confusion_matrix(y_train,y_pred_train))
print('classification report:')
print(classification_report(y_train, y_pred_train))
print('accuracy score  :',accuracy_score(y_train, y_pred_train))
print('roc_auc score   :',roc_auc_score(y_train, y_proba_train))
print('MCC             :',matthews_corrcoef(y_train, y_pred_train))

#Evaluating the Algorithm test set

print('confusion matrix:')
print(confusion_matrix(y_test,y_pred_test))
print('classification report:')
print(classification_report(y_test, y_pred_test))
print('accuracy score  :',accuracy_score(y_test, y_pred_test))
print('roc_auc score   :',roc_auc_score(y_test, y_proba_test))
print('MCC             :',matthews_corrcoef(y_test, y_pred_test))

#plt.plot(fpr1, tpr1, color='red', label='Model 4')
#plt.xlabel('False Positive Rate')
#plt.ylabel('True Positive Rate')
#plt.title('Receiver Operating Characteristic Train Set')
#plt.legend(loc="lower right")
#plt.savefig('ROC AB5 train.png', dpi = 1000)

#plt.plot(fpr2, tpr2, color='black', label='Model 4 - AB 5 (AUC = %0.2f)' % AUC_test)
#plt.xlabel('False Positive Rate')
#plt.ylabel('True Positive Rate')
#plt.title('Receiver Operating Characteristic Test Set')
#plt.legend(loc="lower right")
#plt.savefig('ROC AB5 test.png', dpi = 1000)

#plt.plot(fpr1, tpr1, color='red', label='Model 4 - AB 5 (AUC = %0.2f)' % AUC_train)
#plt.plot(fpr2, tpr2, color='black', label='Model 4 - AB 5 (AUC = %0.2f)' % AUC_test)
#plt.xlabel('False Positive Rate')
#plt.ylabel('True Positive Rate')
#plt.title('Receiver Operating Characteristic Model 4 (5 AB)')
#plt.legend(loc="lower right")
#plt.savefig('ROC AB5.png', dpi = 1000)
