==> 4-SVC_5_AdaBoost_Hyperparameter_Tuning.ipynb <==
import pandas as pd
import pickle
import numpy as np
from sklearn.svm import SVC
from sklearn.datasets import make_classification
from sklearn.ensemble import AdaBoostClassifier
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score

#read train set
train = pd.read_csv('train.csv')
X_train = train.iloc[:,:-1]
y_train = train.iloc[:,-1]
train.shape

#select 5 descriptor
X_train = X_train[['SM1_Dzs','AATS1dv','C2SP2.1','AATSC0v','ATSC2m']]
X_train.head()

#select 5 descriptor
test = pd.read_csv('test.csv')
X_test = test.loc[:,['SM1_Dzs','AATS1dv','C2SP2.1','AATSC0v','ATSC2m']]
y_test = test.iloc[:,-1]
X_test.shape

#define model (parameter)
boost = AdaBoostClassifier(random_state=3)

#svc=SVC(probability=True, kernel='linear')

#parameter dictionary 
gd_param = {'n_estimators':[50, 100, 150, 200, 250],
            'learning_rate':[0.01,0.1,1.0],
            'algorithm':['SAMME.R','SAMME']}

#Pencarian nilai parameter yang ditentukan untuk estimator
gd_search = GridSearchCV(estimator = boost,
                         param_grid = gd_param)

#call the fit method of the class 
modelAB = gd_search.fit(X_train, y_train)

# check the parameters that return the highest accuracy
best_parametersAB = modelAB.best_params_
print(best_parametersAB)

# dump information to that file
pickle.dump(best_parametersAB, open('parameter_AB5.pkl','wb'))

#find the best accuracy achieved
best_result = modelAB.best_score_
print(best_result)
